% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASKFINAL.R
\name{task13_df}
\alias{task13_df}
\title{TASK FINAL - DATA.FRAME VERSION}
\usage{
task13_df(
  integration_file =
    "annotated_GSM3516673_normal_annotated_GSM3516672_tumor_SeuratIntegration.csv",
  annotation_file = "nt_combined_clustering.output.csv"
)
}
\arguments{
\item{integration_file}{Path to annotated integration CSV
(columns: cell, integration_cluster)}

\item{annotation_file}{Path to annotation CSV
(columns: cell, cell_type, sample_type)}
}
\value{
list(merged_df13, count_df13, summary_df13, plot_counts_df13, plot_percent_df13)
}
\description{
Associates cell types to Seurat integration clusters and tissue status (N/T):
it merges integration and annotation data by \code{cell}, returns the merged table,
counts cell types per cluster, creates a summary with normalized percentages
by cluster and sample type (N vs T), and generates bar plots for counts and
normalized \% distributions.
}
